Role

See merge request medeq/api.medeq.ru!544

+++ Modules/Attribute/Http/Controllers/Admin/AttributeController.php
@@ -8,18 +8,20 @@ use Modules\Attribute\Http\Requests\Admin\AttributeCreateRequest;
 use Modules\Attribute\Http\Requests\Admin\AttributeUpdateRequest;
 use Modules\Attribute\Http\Resources\AttributeResource;
 use Modules\Attribute\Models\Attribute;
+use Modules\Attribute\Repositories\AttributeRepository;
 use Modules\Attribute\Services\AttributeStorage;
 
 class AttributeController extends Controller
 {
     public function __construct(
-        protected AttributeStorage $attributeStorage
-    ){
-        $this->authorizeResource(Attribute::class, 'attribute');
-    }
+        protected AttributeStorage $attributeStorage,
+        protected AttributeRepository $attributeRepository
+    ) {}
 
     public function store(AttributeCreateRequest $request)
     {
+        $this->authorize('create', Attribute::class);
+
         $attributeDto = AttributeDto::fromFormRequest($request);
 
         if (!$attributeDto->assigned_by_id) {
@@ -31,15 +33,23 @@ class AttributeController extends Controller
         return new AttributeResource($attribute);
     }
 
-    public function update(Attribute $attribute, AttributeUpdateRequest $request)
+    public function update(int $attribute, AttributeUpdateRequest $request)
     {
+        $attribute = $this->attributeRepository->find($attribute);
+
+        $this->authorize('update', $attribute);
+
         $this->attributeStorage->update($attribute, AttributeDto::fromFormRequest($request));
 
-        return new AttributeResource($attributeModel);
+        return new AttributeResource($attribute);
     }
 
-    public function destroy(Attribute $attribute)
+    public function destroy(int $attribute)
     {
+        $attribute = $this->attributeRepository->find($attribute);
+
+        $this->authorize('delete', $attribute);
+
         $this->attributeStorage->delete($attribute);
 
         return response()->noContent();
