+++ Modules/Case/Http/Controllers/Admin/CaseController.php
@@ -4,6 +4,7 @@ namespace Modules\Case\Http\Controllers\Admin;
 
 
 use App\Http\Controllers\Controller;
+use Modules\Cabinet\Models\Cabinet;
 use Modules\Case\Dto\CaseDto;
 use Modules\Case\Http\Requests\CaseCreateRequest;
 use Modules\Case\Http\Requests\CaseUpdateRequest;
@@ -17,26 +18,34 @@ class CaseController extends Controller
     public function __construct(
         protected CaseStorage $caseStorage,
         protected CaseRepository $caseRepository
-    ) {
-        $this->authorizeResource(CaseModel::class, 'case');
-    }
+    ) {}
 
     public function store(CaseCreateRequest $request)
     {
+        $this->authorize('create', Cabinet::class);
+
         $case = $this->caseStorage->store(CaseDto::fromFormRequest($request));
 
         return new CaseResource($case);
     }
 
-    public function update(CaseModel $case, CaseUpdateRequest $request)
+    public function update(int $case, CaseUpdateRequest $request)
     {
+        $case = $this->caseRepository->find($case);
+
+        $this->authorize('update', $case);
+
         $case = $this->caseStorage->update($case, CaseDto::fromFormRequest($request));
 
         return new CaseResource($case);
     }
 
-    public function destroy(CaseModel $case)
+    public function destroy(int $case)
     {
+        $case = $this->caseRepository->find($case);
+
+        $this->authorize('delete', $case);
+
         $this->caseStorage->delete($case);
 
         return response()->noContent();
