+++ Modules/Category/Http/Controllers/Admin/CategoryController.php
@@ -8,18 +8,20 @@ use Modules\Category\Http\Requests\CategoryCreateRequest;
 use Modules\Category\Http\Requests\CategoryUpdateRequest;
 use Modules\Category\Http\Resources\CategoryResource;
 use Modules\Category\Models\Category;
+use Modules\Category\Repositories\CategoryRepository;
 use Modules\Category\Services\CategoryStorage;
 
 class CategoryController extends Controller
 {
     public function __construct(
-        protected CategoryStorage $categoryStorage
-    ) {
-        $this->authorizeResource(Category::class, 'category');
-    }
+        protected CategoryStorage $categoryStorage,
+        protected CategoryRepository $categoryRepository
+    ) {}
 
     public function store(CategoryCreateRequest $request)
     {
+        $this->authorize('create', Category::class);
+
         $categoryDto = CategoryDto::fromFormRequest($request);
 
         if (!$categoryDto->assigned_by_id) {
@@ -31,8 +33,12 @@ class CategoryController extends Controller
         return new CategoryResource($category);
     }
 
-    public function update(Category $category, CategoryUpdateRequest $request)
+    public function update(int $category, CategoryUpdateRequest $request)
     {
+        $category = $this->categoryRepository->find($category);
+
+        $this->authorize('update', $category);
+
         $category = $this->categoryStorage->update(
             $category, CategoryDto::fromFormRequest($request)
         );
@@ -40,8 +46,12 @@ class CategoryController extends Controller
         return new CategoryResource($category);
     }
 
-    public function destroy(Category $category)
+    public function destroy(int $category)
     {
+        $category = $this->categoryRepository->find($category);
+
+        $this->authorize('delete', $category);
+
         $this->categoryStorage->delete($category);
 
         return response()->noContent();
