+++ Modules/Customer/Http/Controllers/Admin/CustomerReviewController.php
@@ -9,18 +9,20 @@ use Modules\Customer\Http\Requests\Admin\CustomerReviewCreateRequest;
 use Modules\Customer\Http\Requests\Admin\CustomerReviewUpdateRequest;
 use Modules\Customer\Http\Resources\CustomerReviewResource;
 use Modules\Customer\Models\CustomerReview;
+use Modules\Customer\Repositories\CustomerReviewRepository;
 use Modules\Customer\Services\Admin\CustomerReviewStorage;
 
 class CustomerReviewController extends Controller
 {
     public function __construct(
-        protected CustomerReviewStorage $storage
-    ) {
-        $this->authorizeResource(CustomerReview::class, 'customer_review');
-    }
+        protected CustomerReviewStorage $storage,
+        protected CustomerReviewRepository $customerReviewRepository
+    ) {}
 
     public function store(CustomerReviewCreateRequest $request): CustomerReviewResource
     {
+        $this->authorize('create', CustomerReview::class);
+
         $customerReview = $this->storage->store(
             CustomerReviewDto::fromFormRequest($request)
         );
@@ -28,8 +30,12 @@ class CustomerReviewController extends Controller
         return new CustomerReviewResource($customerReview);
     }
 
-    public function update(CustomerReviewUpdateRequest $request, CustomerReview $customer_review): CustomerReviewResource
+    public function update(CustomerReviewUpdateRequest $request, int $customer_review): CustomerReviewResource
     {
+        $customer_review = $this->customerReviewRepository->find($customer_review);
+
+        $this->authorize('update', $customer_review);
+
         $this->storage->update(
             $customer_review,
             CustomerReviewDto::fromFormRequest($request)
@@ -38,8 +44,12 @@ class CustomerReviewController extends Controller
         return new CustomerReviewResource($customer_review);
     }
 
-    public function destroy(CustomerReview $customer_review): Response
+    public function destroy(int $customer_review): Response
     {
+        $customer_review = $this->customerReviewRepository->find($customer_review);
+
+        $this->authorize('delete', $customer_review);
+
         $this->storage->delete($customer_review);
 
         return response()->noContent();
