+++ Modules/Page/Http/Controllers/Admin/PageController.php
@@ -8,18 +8,20 @@ use Modules\Page\Http\Requests\PageCreateRequest;
 use Modules\Page\Http\Requests\PageUpdateRequest;
 use Modules\Page\Http\Resources\PageResource;
 use Modules\Page\Models\Page;
+use Modules\Page\Repositories\PageRepository;
 use Modules\Page\Services\PageStorage;
 
 class PageController extends Controller
 {
     public function __construct(
-        protected PageStorage $pageStorage
-    ) {
-        $this->authorizeResource(Page::class, 'page');
-    }
+        protected PageStorage $pageStorage,
+        protected PageRepository $pageRepository
+    ) {}
 
     public function store(PageCreateRequest $request)
     {
+        $this->authorize('create', Page::class);
+
         $pageDto = PageDto::fromFormRequest($request);
 
         if (!$pageDto->assigned_by_id) {
@@ -31,15 +33,23 @@ class PageController extends Controller
         return new PageResource($pageModel);
     }
 
-    public function update(Page $page, PageUpdateRequest $request)
+    public function update(int $page, PageUpdateRequest $request)
     {
+        $page = $this->pageRepository->find($page);
+
+        $this->authorize('update', $page);
+
         $page = $this->pageStorage->update($page, PageDto::fromFormRequest($request));
 
         return new PageResource($page);
     }
 
-    public function destroy(Page $page)
+    public function destroy(int $page)
     {
+        $page = $this->pageRepository->find($page);
+
+        $this->authorize('delete', $page);
+
         $this->pageStorage->delete($page);
 
         return response()->noContent();
